+++ src/components/Register/Register.scss
@@ -1,7 +1,6 @@
 @import "./../../styles/variables";
 @import "./../../styles/constants";
-
-.login-form {
+.registration-form {
   align-items: center;
   display: flex;
   flex-direction: column;
@@ -18,7 +17,6 @@
 
   &__field {
     margin-bottom: 1rem;
-    position: relative;
   }
   &__label {
     display: block;
@@ -27,14 +25,8 @@
   }
   &__error-text {
     color: $error-color;
-    position: absolute;
-    top: 0;
-    left: 5.2rem;
   }
-  &__error-message {
-    color: $error-color;
-    margin-top: 1rem;
-  }
+
   &__footer-link {
     margin-top: 1rem;
   }
@@ -52,6 +44,7 @@
     border: none;
     color: $primary-color;
     cursor: pointer;
+    font-weight: $font-weight-black;
     padding: 0.5rem 1rem;
   }
 }
