+++ src/constants/actionTypes.js
@@ -1,4 +1,3 @@
 export const REGISTER_USER = "REGISTER_USER";
 export const LOGIN_USER = "LOGIN_USER";
 export const LOGOUT_USER = "LOGOUT_USER";
-export const ADD_BLOG_POST = "ADD_BLOG_POST";
