+++ src/main.jsx
@@ -1,8 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import Home from "./pages/Home.jsx";
-import Login from "./pages/Login.jsx";
-import Register from "./pages/Register.jsx";
 import Layout from "./Layout.jsx";
 import "./index.scss";
 import {
@@ -13,6 +11,8 @@ import {
 } from "react-router-dom";
 import { Provider } from "react-redux";
 import { store } from "./store/store.js";
+import Login from "./components/Login/Login.jsx";
+import Register from "./components/Register/Register.jsx";
 
 const router = createBrowserRouter(
   createRoutesFromElements(
