--- src/pages/Login.jsx
@@ -1,16 +1,0 @@
-import React, { useState } from "react";
-import LoginForm from "../components/LoginForm/LoginForm";
-
-const Login = () => {
-  const [errorMessage, setErrorMessage] = useState("");
-  return (
-    <>
-      <LoginForm
-        errorMessage={errorMessage}
-        setErrorMessage={setErrorMessage}
-      />
-    </>
-  );
-};
-
-export default Login;
