--- src/pages/Register.jsx
@@ -1,12 +1,0 @@
-import React from "react";
-import RegistrationForm from "../components/RegistrationForm/RegistrationForm";
-
-const Register = () => {
-  return (
-    <>
-      <RegistrationForm />
-    </>
-  );
-};
-
-export default Register;
